# openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Todo API
  description: A very simple todo API
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /todo/:
    get:
      security:
        - bearerAuth: []
      description: Get a user's todo list
      
      responses:
        201:
          description: Todo Retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoGetArray'  
        
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
                    
    post:
      security:
        - bearerAuth: []
      description: Create a todo for a user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'    
      
      responses:
        201:
          description: Todo Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'  
        
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:   
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
...
    Todo:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: test
        description:
          type: string
          example: My test todo
    TodoGetArray:
      type: array
      items:
        properties:
          title:
            type: string
            example: test
          description:
            type: string
            example: My test todo
          code:
            type: string
            example: jfuh-8yhsh-sj44-sjs4
          done:
            type: boolean
            example: False
          date_created:
            type: string
            example: 2021-07-0713:00:18.259396
          user:
            type: object
            properties:
              email:
                type: string
                example: test@test.com
              first_name:  
                type: string
                example: Test
              last_name:
                type: string
                example: Test
...